Backend/requirements.txt
fastapi==0.115.9
uvicorn[standard]==0.23.2
psycopg2-binary==2.9.9
qdrant-client==1.15.1
langchain==0.3.25
langchain-google-genai==2.0.10
langchain-community==0.3.24
python-dotenv==1.0.1
requests==2.31.0
numpy==1.26.2
joblib==1.4.2
gpt4all==2.8.2
sqlalchemy==2.0.41
pydantic==2.11.5
scikit-learn==1.5.1
